pw_boot_cortex_m: Ensure FPCCR.LSPACT is cleared
Cortex-M processors support Lazy floating-point state preservation,
which aims to reduce exception latency by only pushing FPU register
state on exception entry when strictly necessary. This is tracked by
the processor in the FPCCR.LSPACT bit, which indicates that FPU
registers have not been pushed to the stack yet.

If the previous boot stage left this feature active, but did not finish
the exception completely which caused the state save, pw_boot_Entry
will be running with FPCCR.LSPACT=1. This is problematic, as the very
first FPU instruction to be executed afterwards will invoke an implicit
FPU state restore by pushing the preserved FPU state under the address
found in FPCAR. This results in a random portion of memory being
overwritten once an FPU instruction is executed. This can happen very
early, during static constructor initialization, due to the usage of
FPU extension registers for optimized memory loads/stores.

This explicitly clears the LSPACT bit to invalidate the lazy floating
point state, so the first FPU instruction to be executed does not
silently corrupt random portions of memory.

Change-Id: Idafa0e62ea1475f7cd5c49bbf062c28149bfee13