pw_flash: Add Read/GetFlashParameters/GetPageLayout
This adds flash reading to the pw::flash interface, as well as the
ability to introspect the physical page layout of the flash device.
This is required, as erase/write operations must respect the underlying
flash constraints, and with this it is possible for the application to
adjust according to the used flash device.

Change-Id: I95d31f44e2b80a2a6192b7d263eb949118ae6424